requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.41
joblib==1.5.3
jpype1==1.6.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
orhelper==0.1.3
packaging==26.0
pillow==12.1.1
pyaml==26.2.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scikit-optimize==0.10.2
scipy==1.17.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
threadpoolctl==3.6.0
